+++ projects/_03_independent_cqc/_03_impute/fargate/impute_ind_cqc_ascwds_and_pir.py
@@ -1,6 +1,9 @@
 import polars as pl
 
 from polars_utils import utils
+from projects._03_independent_cqc._03_impute.fargate.utils.convert_pir_people_to_filled_posts import (
+    convert_pir_to_filled_posts,
+)
 from utils.column_names.ind_cqc_pipeline_columns import IndCqcColumns as IndCQC
 
 
@@ -24,7 +27,7 @@ def main(
 
     # model_primary_service_rate_of_change_trendline - ascwds_rate_of_change_trendline_model
 
-    # model_pir_filled_posts
+    lf = convert_pir_to_filled_posts(lf)
 
     # merge_ascwds_and_pir_filled_post_submissions