+++ projects/_03_independent_cqc/_03_impute/tests/fargate/test_impute_ind_cqc_ascwds_and_pir.py
@@ -25,10 +25,12 @@ class ImputeIndCqcAscwdsAndPirTests(unittest.TestCase):
     mock_data = Mock(name="data")
 
     @patch(f"{PATCH_PATH}.utils.sink_to_parquet")
+    @patch(f"{PATCH_PATH}.convert_pir_to_filled_posts")
     @patch(f"{PATCH_PATH}.utils.scan_parquet", return_value=mock_data)
     def test_main_runs_successfully(
         self,
         scan_parquet_mock: Mock,
+        convert_pir_to_filled_posts_mock: Mock,
         sink_to_parquet_mock: Mock,
     ):
 
@@ -38,6 +40,7 @@ class ImputeIndCqcAscwdsAndPirTests(unittest.TestCase):
         )
 
         scan_parquet_mock.assert_called_once()
+        convert_pir_to_filled_posts_mock.assert_called_once()
         sink_to_parquet_mock.assert_called_once_with(
             ANY,
             self.TEST_DESTINATION,