+++ projects/_03_independent_cqc/unittest_data/polars_ind_cqc_test_file_schemas.py
@@ -25,6 +25,9 @@ from utils.column_names.ind_cqc_pipeline_columns import (
 )
 from utils.column_names.ind_cqc_pipeline_columns import IndCqcColumns as IndCQC
 from utils.column_names.ind_cqc_pipeline_columns import (
+    NonResWithAndWithoutDormancyCombinedColumns as NRModel_TempCol,
+)
+from utils.column_names.ind_cqc_pipeline_columns import (
     NullGroupedProviderColumns as NGPcol,
 )
 
@@ -1199,6 +1202,20 @@ class WinsorizeCareHomeFilledPostsPerBedRatioOutliersSchema:
 
 
 @dataclass
+class ConvertPirPeopleToFilledPostsSchema:
+    input_schema = pl.Schema(
+        [
+            (IndCQC.care_home, pl.String()),
+            (IndCQC.pir_people_directly_employed_dedup, pl.Float64()),
+            (IndCQC.ascwds_filled_posts_dedup_clean, pl.Float64()),
+        ]
+    )
+    output_schema = pl.Schema(
+        list(input_schema.items()) + [(IndCQC.pir_filled_posts_model, pl.Float64())]
+    )
+
+
+@dataclass
 class NullCtPostsToBedsOutliers:
     null_ct_posts_to_beds_outliers_schema = pl.Schema(
         [
@@ -1316,5 +1333,53 @@ class EstimateFilledPostsModelsUtils:
             **join_ind_cqc_schema,
             join_test_model: pl.Float64,
             f"{join_test_model}_run_id": pl.String,
+        }
+    )
+
+
+@dataclass
+class ModelNonResWithAndWithoutDormancyCombinedSchemas:
+    estimated_posts_schema = pl.Schema(
+        {
+            IndCQC.location_id: pl.String,
+            IndCQC.cqc_location_import_date: pl.Date,
+            IndCQC.care_home: pl.String,
+            IndCQC.related_location: pl.String,
+            IndCQC.time_registered: pl.Int64,
+            IndCQC.non_res_without_dormancy_model: pl.Float64,
+            IndCQC.non_res_with_dormancy_model: pl.Float64,
+            IndCQC.non_res_combined_model: pl.Float64,
+        }
+    )
+
+    group_time_registered_to_six_month_bands_schema = pl.Schema(
+        {
+            IndCQC.location_id: pl.String,
+            IndCQC.time_registered: pl.Int64,
+            NRModel_TempCol.time_registered_banded_and_capped: pl.Float64,
+        }
+    )
+
+    calculate_and_apply_model_ratios_schema = pl.Schema(
+        {
+            IndCQC.related_location: pl.String,
+            NRModel_TempCol.time_registered_banded_and_capped: pl.Int64,
+            IndCQC.non_res_without_dormancy_model: pl.Float64,
+            IndCQC.non_res_with_dormancy_model: pl.Float64,
+            NRModel_TempCol.avg_with_dormancy: pl.Float64,
+            NRModel_TempCol.avg_without_dormancy: pl.Float64,
+            NRModel_TempCol.adjustment_ratio: pl.Float64,
+            NRModel_TempCol.non_res_without_dormancy_model_adjusted: pl.Float64,
+        }
+    )
+
+    calculate_and_apply_residuals_schema = pl.Schema(
+        {
+            IndCQC.location_id: pl.String,
+            IndCQC.cqc_location_import_date: pl.Date,
+            IndCQC.non_res_with_dormancy_model: pl.Float64,
+            NRModel_TempCol.non_res_without_dormancy_model_adjusted: pl.Float64,
+            NRModel_TempCol.residual_at_overlap: pl.Float64,
+            NRModel_TempCol.non_res_without_dormancy_model_adjusted_and_residual_applied: pl.Float64,
         }
     )